Added: Mapbox and info window
