+++ app/assets/stylesheets/application.scss
@@ -12,3 +12,4 @@
 @import "components/index";
 @import "pages/index";
 @import "components/card";
+@import "components/map";
