+++ app/assets/stylesheets/components/_footer.scss
@@ -1,10 +1,9 @@
 .footer {
   background: white;
   display: flex;
-  align-items: center;
-  justify-content: space-between;
+  justify-content: center;
   height: 100px;
-  padding: 0px 50px;
+  width: 100vw;
   color: rgba(0,0,0,0.3);
 }
 .footer-links {
