+++ app/javascript/plugins/init_autocomplete.js
@@ -1,0 +1,10 @@
+import places from 'places.js';
+
+const initAutocomplete = () => {
+  const addressInput = document.getElementById('flat_address');
+  if (addressInput) {
+    places({ container: addressInput });
+  }
+};
+
+export { initAutocomplete };
